Extract generic ssz types from p2p package to sszutil
