+++ shared/sszutil/BUILD.bazel
@@ -3,10 +3,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["deep_equal.go"],
+    srcs = [
+        "deep_equal.go",
+        "types.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/shared/sszutil",
     visibility = ["//visibility:public"],
     deps = [
+        "//shared/htrutils:go_default_library",
+        "@com_github_ferranbt_fastssz//:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@org_golang_google_protobuf//proto:go_default_library",
     ],
@@ -15,11 +21,15 @@ go_library(
 go_test(
     name = "go_default_test",
     size = "small",
-    srcs = ["deep_equal_test.go"],
+    srcs = [
+        "deep_equal_test.go",
+        "types_test.go",
+    ],
     deps = [
         ":go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/testutil/assert:go_default_library",
+        "//shared/testutil/require:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
     ],
 )
