+++ slasher/rpc/server_test.go
@@ -10,13 +10,13 @@ import (
 	"github.com/prysmaticlabs/eth2-types"
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	p2ptypes "github.com/prysmaticlabs/prysm/beacon-chain/p2p/types"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state"
 	"github.com/prysmaticlabs/prysm/shared/bls"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/mock"
 	"github.com/prysmaticlabs/prysm/shared/p2putils"
 	"github.com/prysmaticlabs/prysm/shared/params"
+	"github.com/prysmaticlabs/prysm/shared/sszutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
@@ -237,7 +237,7 @@ func TestServer_IsSlashableBlock(t *testing.T) {
 			sbbh.Header.BodyRoot = bytesutil.PadTo([]byte(fmt.Sprintf("%d", j)), 32)
 			bhr, err := sbbh.Header.HashTreeRoot()
 			assert.NoError(t, err)
-			sszBytes := p2ptypes.SSZBytes(bhr[:])
+			sszBytes := sszutil.SSZBytes(bhr[:])
 			root, err := helpers.ComputeSigningRoot(&sszBytes, domain)
 			assert.NoError(t, err)
 			sbbh.Signature = keys[sbbh.Header.ProposerIndex].Sign(root[:]).Marshal()
@@ -306,7 +306,7 @@ func TestServer_IsSlashableBlockNoUpdate(t *testing.T) {
 	require.NoError(t, err)
 	bhr, err := savedBlock.Header.HashTreeRoot()
 	require.NoError(t, err)
-	sszBytes := p2ptypes.SSZBytes(bhr[:])
+	sszBytes := sszutil.SSZBytes(bhr[:])
 	root, err := helpers.ComputeSigningRoot(&sszBytes, domain)
 	require.NoError(t, err)
 	blockSig := keys[savedBlock.Header.ProposerIndex].Sign(root[:])
