+++ validator/client/BUILD.bazel
@@ -22,7 +22,6 @@ go_library(
     visibility = ["//validator:__subpackages__"],
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
-        "//beacon-chain/p2p/types:go_default_library",
         "//proto/validator/accounts/v2:go_default_library",
         "//shared/blockutil:go_default_library",
         "//shared/bls:go_default_library",
@@ -37,6 +36,7 @@ go_library(
         "//shared/rand:go_default_library",
         "//shared/slashutil:go_default_library",
         "//shared/slotutil:go_default_library",
+        "//shared/sszutil:go_default_library",
         "//shared/timeutils:go_default_library",
         "//shared/traceutil:go_default_library",
         "//validator/accounts/wallet:go_default_library",
